samples/microsoft-agent-framework/single-agent/pyproject.toml
[project]
name = "maf-single-agent"
version = "0.1.0"
description = "Microsoft Agent Framework single-agent document analysis-summary-review pipeline"
requires-python = ">=3.11"

dependencies = [
    "azure-ai-projects>=1.0.0b7",
    "azure-ai-agents>=1.0.0b4",
    "azure-identity>=1.15.0",
    "aiohttp>=3.9.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
shared = { workspace = true }
